requirements.txt
matplotlib==3.2.2
numpy==1.24.1
Pillow==9.4.0
scikit_learn==1.2.0
scipy==1.5.0
torch==1.12.1
torchvision==0.13.1
